+++ xill-cli/src/test/java/nl/xillio/xill/cli/XillCLITest.java
@@ -16,14 +16,21 @@
 package nl.xillio.xill.cli;
 
 import nl.xillio.xill.XillProcessor;
+import org.apache.commons.io.FileUtils;
 import org.mockito.ArgumentCaptor;
 import org.testng.annotations.Test;
 
 import java.io.ByteArrayOutputStream;
+import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.PrintStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.nio.file.StandardOpenOption;
+import java.util.Collections;
+import java.util.zip.ZipEntry;
+import java.util.zip.ZipOutputStream;
 import java.nio.file.StandardOpenOption;
 import java.util.Collections;
 
@@ -31,7 +38,6 @@ import static org.mockito.Matchers.anyString;
 import static org.mockito.Mockito.*;
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertTrue;
-
 
 public class XillCLITest {
 
@@ -137,7 +143,7 @@ public class XillCLITest {
 
         // Run by path
         XillCLI xillCLI = new XillCLI();
-        xillCLI.setArgs(new String[]{"fqn.Path", "-p", projectDir.toString()});
+        xillCLI.setArgs(new String[]{"fqn.Path", "-w", projectDir.toString()});
 
         // Validate
         ProgramReturnCode returnCode = xillCLI.run();
@@ -147,5 +153,48 @@ public class XillCLITest {
         Files.delete(robotFile);
         Files.delete(robotFile.getParent());
         Files.delete(projectDir);
+    }
+
+    @Test
+    public void testRunRobotWithIncludes() throws IOException {
+        // Create a folder with a robot.
+        Path tempDir = Files.createTempDirectory("unit-test");
+        Path projectDir = tempDir.resolve("project");
+        Path includedDir = tempDir.resolve("includedDir");
+        Files.createDirectories(projectDir);
+        Files.createDirectories(includedDir);
+
+        // Create an archive with a robot.
+        Path includedArchive = tempDir.resolve("archive.xip");
+        ZipOutputStream archiveStream = new ZipOutputStream(new FileOutputStream(includedArchive.toFile()));
+        ZipEntry entry = new ZipEntry("robots/ArchiveRobot.xill");
+        archiveStream.putNextEntry(entry);
+        archiveStream.write("use System; function archive() { System.print('This is an included robot from an archive.'); }".getBytes());
+        archiveStream.closeEntry();
+        archiveStream.close();
+
+        // The project robots.
+        Path baseRobot = projectDir.resolve("BaseRobot.xill");
+        Files.write(baseRobot, Collections.singletonList(
+                "use System;" +
+                        "include ProjectRobot; include DirectoryRobot; include ArchiveRobot;" +
+                        "System.print('This is the base robot.');" +
+                        "project(); directory(); archive();"
+        ), StandardOpenOption.CREATE);
+        Path projectRobot = projectDir.resolve("ProjectRobot.xill");
+        Files.write(projectRobot, Collections.singletonList("use System; function project() { System.print('This is another project robot.'); }"), StandardOpenOption.CREATE);
+
+        // The included directory robot.
+        Path directoryRobot = includedDir.resolve("DirectoryRobot.xill");
+        Files.write(directoryRobot, Collections.singletonList("use System; function directory() { System.print('This is an included robot from a directory.'); }"), StandardOpenOption.CREATE);
+
+        // Run the robot.
+        XillCLI xillCLI = new XillCLI();
+        xillCLI.setArgs(new String[]{"BaseRobot", "-w", projectDir.toString(), "-r", includedDir + ":" + includedArchive});
+        ProgramReturnCode returnCode = xillCLI.run();
+        assertEquals(returnCode, ProgramReturnCode.OK);
+
+        // Delete project
+        FileUtils.forceDeleteOnExit(tempDir.toFile());
     }
 }