+++ xill-processor/src/main/java/nl/xillio/xill/plugins/system/services/info/RobotRuntimeInfo.java
@@ -16,21 +16,27 @@
 package nl.xillio.xill.plugins.system.services.info;
 
 import com.google.common.util.concurrent.Service;
+import nl.xillio.xill.api.XillEnvironment;
+import nl.xillio.xill.api.components.RobotID;
 import nl.xillio.xill.api.construct.ConstructContext;
 import nl.xillio.xill.services.PropertiesProvider;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
+import java.util.regex.Pattern;
 
 /**
  * This class represents a {@link Service} that provides information about the currently running robot
  */
 public class RobotRuntimeInfo implements PropertiesProvider {
+    private static final Pattern FQN_PATTERN = Pattern.compile("[A-z_][0-9A-z_]*(\\.[A-z_][0-9A-z_]*)*");
 
     private final Map<String, Object> properties = new LinkedHashMap<>();
-    private final String robotPath;
-    private final String rootRobotPath;
-    private final String projectPath;
+    private final String robotUrl;
+    private final String rootRobotUrl;
+    private final String workingDirectory;
+    private final String resourcePath;
+    private final String qualifiedName;
 
     /**
      * Create a new {@link RobotRuntimeInfo} for a {@link ConstructContext}
@@ -38,13 +44,28 @@ public class RobotRuntimeInfo implements PropertiesProvider {
      * @param context the context
      */
     public RobotRuntimeInfo(final ConstructContext context) {
-        robotPath = context.getRobotID().getPath().getAbsolutePath();
-        rootRobotPath = context.getRootRobot().getPath().getAbsolutePath();
-        projectPath = context.getRobotID().getProjectPath().getAbsolutePath();
+        robotUrl = context.getRobotID().getURL().toString();
+        rootRobotUrl = context.getRootRobot().getURL().toString();
+        workingDirectory = context.getWorkingDirectory().toString();
+        resourcePath = context.getRobotID().getResourcePath();
+        qualifiedName = getQualifiedName(context.getRobotID());
 
-        properties.put("robotPath", robotPath);
-        properties.put("rootRobotPath", rootRobotPath);
-        properties.put("projectPath", projectPath);
+        properties.put("robotUrl", robotUrl);
+        properties.put("resourcePath", resourcePath);
+        properties.put("qualifiedName", qualifiedName);
+        properties.put("rootRobotUrl", rootRobotUrl);
+        properties.put("projectPath", workingDirectory);
+    }
+
+    private String getQualifiedName(RobotID robotID) {
+        String result = robotID.getResourcePath().replaceAll("[/\\\\]", ".");
+        if (result.endsWith(XillEnvironment.ROBOT_EXTENSION)) {
+            result = result.substring(0, result.length() - XillEnvironment.ROBOT_EXTENSION.length());
+        }
+        if (FQN_PATTERN.matcher(result).matches()) {
+            return result;
+        }
+        return null;
     }
 
     @Override
@@ -55,22 +76,36 @@ public class RobotRuntimeInfo implements PropertiesProvider {
     /**
      * @return the robotPath
      */
-    public String getRobotPath() {
-        return robotPath;
+    public String getRobotUrl() {
+        return robotUrl;
     }
 
     /**
      * @return the rootRobotPath
      */
-    public String getRootRobotPath() {
-        return rootRobotPath;
+    public String getRootRobotUrl() {
+        return rootRobotUrl;
+    }
+
+    /**
+     * @return the resource path
+     */
+    public String getResourcePath() {
+        return resourcePath;
+    }
+
+    /**
+     * @return the robot's qualified name or null
+     */
+    public String getQualifiedName() {
+        return qualifiedName;
     }
 
     /**
      * @return the projectPath
      */
-    public String getProjectPath() {
-        return projectPath;
+    public String getWorkingDirectory() {
+        return workingDirectory;
     }
 
 }